Move to Hosted vs Timed service

+++ WeatherForecastMvc/HostedServices/ForecastCleanuperService.cs
@@ -1,6 +1,7 @@
 internal interface IForecastCleanuperService
 {
-    Task DoWork(CancellationToken stoppingToken);
+    Task CleanupPeriodically(uint cleanupPeriodSec, CancellationToken stoppingToken);
+    void CleanupOnce();
 }
 
 internal class ForecastCleanuperService : IForecastCleanuperService
@@ -8,34 +9,50 @@ internal class ForecastCleanuperService : IForecastCleanuperService
     private int executionCount = 0;
     private readonly ILogger _logger;
     private readonly ForecastContext _context;
+    private readonly IConfiguration Configuration;
 
-    public ForecastCleanuperService(ILogger<ForecastCleanuperService> logger, ForecastContext context)
+    public ForecastCleanuperService(ILogger<ForecastCleanuperService> logger, ForecastContext context, IConfiguration configuration)
     {
         _logger = logger;
         _context = context;
+        Configuration = configuration;
     }
 
-    public async Task DoWork(CancellationToken stoppingToken)
+    public async Task CleanupPeriodically(uint cleanupPeriodSec, CancellationToken stoppingToken)
     {
         while (!stoppingToken.IsCancellationRequested)
         {
             executionCount++;
             _logger.LogInformation("ForecastCleanuperService is working. Count: {Count}", executionCount);
-            CleanupOldRecords(5);
-            await Task.Delay(TimeSpan.FromSeconds(10), stoppingToken);
+            CleanupOldRecords();
+            await Task.Delay(TimeSpan.FromSeconds(cleanupPeriodSec), stoppingToken);
         }
     }
 
-    private void CleanupOldRecords(uint calendarSpanDays)
+    public void CleanupOnce()
     {
-        _logger.LogInformation($"Cleaning up old records. Time:{DateTime.Now.ToShortTimeString()}");
+        CleanupOldRecords();
+    }
+
+    private void CleanupOldRecords()
+    {
+        _logger.LogInformation($"Cleaning up old records. Time:{DateTime.Now.ToLongTimeString()}");
+
+        if (!uint.TryParse(Configuration["WeatherForecastMvcConfig:CalendarSpanDays"], out var oldRecordsCleanupDays))
+        {
+            oldRecordsCleanupDays = 10;
+        }
 
         var oldRecords = _context.DayForecast
-        .Where(f => f.Date < DateTime.Today.AddDays(-calendarSpanDays))
-        .ToArray();
+        .Where(f => f.Date < DateTime.Today.AddDays(-oldRecordsCleanupDays));
 
-        string dates = String.Join(',', oldRecords.Select(r => r.Date));
-        _logger.LogInformation($"Removing records for dates:{dates}");
+        if (!oldRecords.Any())
+        {
+            return;
+        }
+
+        string dates = String.Join(',', oldRecords.Select(r => r.Date.ToShortDateString()));
+        _logger.LogWarning($"Removing records for dates:{dates}");
 
         _context.DayForecast.RemoveRange(oldRecords);
         _context.SaveChanges();
